# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "House Price Prediction API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/policies/feature/simulate": {
      "post": {
        "tags": [
          "policies"
        ],
        "summary": "Simulate Feature Policies",
        "operationId": "simulate_feature_policies_v1_policies_feature_simulate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FeaturePolicySimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeaturePolicySimulationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/predictions": {
      "post": {
        "tags": [
          "predictions"
        ],
        "summary": "Create Prediction",
        "operationId": "create_prediction_v1_predictions_post",
        "parameters": [
          {
            "name": "x-correlation-id",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Correlation-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PredictionRequestPayload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "List Predictions",
        "operationId": "list_predictions_v1_predictions_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/predictions/{prediction_id}": {
      "get": {
        "tags": [
          "predictions"
        ],
        "summary": "Get Prediction",
        "operationId": "get_prediction_v1_predictions__prediction_id__get",
        "parameters": [
          {
            "name": "prediction_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Prediction Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/properties/normalize": {
      "post": {
        "tags": [
          "properties"
        ],
        "summary": "Normalize Address",
        "operationId": "normalize_address_v1_properties_normalize_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddressPayload"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NormalizedAddress"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/validation/run-scenario-batch": {
      "post": {
        "tags": [
          "validation"
        ],
        "summary": "Run Scenario Batch",
        "description": "Run one or more registered scenarios through the full audit pipeline.\n\nPass ``scenario_ids`` to select a subset; omit (or send ``null``) to run\nall registered scenarios.",
        "operationId": "run_scenario_batch_v1_validation_run_scenario_batch_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioBatchPipelineRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioBatchPipelineResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AddressPayload": {
        "properties": {
          "address_line_1": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Address Line 1"
          },
          "address_line_2": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line 2"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "City"
          },
          "state": {
            "type": "string",
            "maxLength": 30,
            "minLength": 2,
            "title": "State"
          },
          "postal_code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Postal Code"
          },
          "country": {
            "type": "string",
            "maxLength": 60,
            "minLength": 2,
            "title": "Country",
            "default": "US"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "address_line_1",
          "city",
          "state",
          "postal_code"
        ],
        "title": "AddressPayload"
      },
...
      "FeaturePolicySimulationRequest": {
        "properties": {
          "address_line_1": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Address Line 1"
          },
          "address_line_2": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line 2"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "City"
          },
          "state": {
            "type": "string",
            "maxLength": 30,
            "minLength": 2,
            "title": "State"
          },
          "postal_code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Postal Code"
          },
          "country": {
            "type": "string",
            "maxLength": 60,
            "minLength": 2,
            "title": "Country",
            "default": "US"
          },
          "policy_names": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Policy Names"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "address_line_1",
          "city",
          "state",
          "postal_code"
        ],
        "title": "FeaturePolicySimulationRequest"
      },
      "FeaturePolicySimulationResponse": {
        "properties": {
          "normalized_address": {
            "$ref": "#/components/schemas/NormalizedAddress"
          },
          "provider_name": {
            "type": "string",
            "title": "Provider Name"
          },
          "simulations": {
            "items": {
              "$ref": "#/components/schemas/FeaturePolicySimulationItem"
            },
            "type": "array",
            "title": "Simulations"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "normalized_address",
          "provider_name",
          "simulations"
        ],
        "title": "FeaturePolicySimulationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictionDetailResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "prediction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Prediction Id"
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Correlation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "received",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "predicted_price": {
            "type": "number",
            "title": "Predicted Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USD"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "feature_snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Feature Snapshot Id"
          },
          "normalized_address": {
            "$ref": "#/components/schemas/NormalizedAddress"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "was_reused": {
            "type": "boolean",
            "title": "Was Reused",
            "default": false
          },
          "source_prediction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Prediction Id"
          },
          "selected_feature_policy_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Feature Policy Name"
          },
          "selected_feature_policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Feature Policy Version"
          },
          "key_features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Key Features",
            "description": "Key property features used for this prediction, e.g. bedrooms, living area sq ft, lot area, bathrooms, year built. Only populated features (non-null) are included."
          },
          "feature_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Source",
            "description": "Data source used to build the feature vector: 'census_context', 'census_context_with_backfill', 'heuristic', or 'fake'. Useful for auditing data quality in production."
          },
          "feature_snapshot": {
            "$ref": "#/components/schemas/FeatureSnapshotSummary"
          },
          "provider_responses": {
            "items": {
              "$ref": "#/components/schemas/ProviderResponseSummary"
            },
            "type": "array",
            "title": "Provider Responses"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "prediction_id",
          "correlation_id",
          "status",
          "predicted_price",
          "model_name",
          "model_version",
          "feature_snapshot_id",
          "normalized_address",
          "submitted_at",
          "generated_at",
          "feature_snapshot",
          "provider_responses"
        ],
        "title": "PredictionDetailResponse"
      },
...
      "PredictionListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PredictionListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "default": 0
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "default": 10
          },
          "offset": {
            "type": "integer",
            "title": "Offset",
            "default": 0
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "items"
        ],
        "title": "PredictionListResponse"
      },
      "PredictionRequestPayload": {
        "properties": {
          "address_line_1": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Address Line 1"
          },
          "address_line_2": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Address Line 2"
          },
          "city": {
            "type": "string",
            "maxLength": 80,
            "minLength": 1,
            "title": "City"
          },
          "state": {
            "type": "string",
            "maxLength": 30,
            "minLength": 2,
            "title": "State"
          },
          "postal_code": {
            "type": "string",
            "maxLength": 20,
            "minLength": 3,
            "title": "Postal Code"
          },
          "country": {
            "type": "string",
            "maxLength": 60,
            "minLength": 2,
            "title": "Country",
            "default": "US"
          },
          "requested_by": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ],
            "title": "Requested By"
          },
          "preferred_policy_name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 80
              },
              {
                "type": "null"
              }
            ],
            "title": "Preferred Policy Name",
            "description": "Optional feature policy to apply for this specific request. When set, overrides the server-configured default and any state-level policy routing. Must match a known policy name from GET /v1/policies/feature."
          },
          "feature_overrides": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Overrides",
            "description": "Optional map of known feature values to inject before feature assembly. Keys must match expected model feature names (e.g. BedroomAbvGr, GrLivArea). Overrides provider-fetched values for matching keys. Prediction reuse is disabled whenever this field is non-null."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "address_line_1",
          "city",
          "state",
          "postal_code"
        ],
        "title": "PredictionRequestPayload"
      },
      "PredictionResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid",
            "title": "Request Id"
          },
          "prediction_id": {
            "type": "string",
            "format": "uuid",
            "title": "Prediction Id"
          },
          "correlation_id": {
            "type": "string",
            "format": "uuid",
            "title": "Correlation Id"
          },
          "status": {
            "type": "string",
            "enum": [
              "received",
              "completed",
              "failed"
            ],
            "title": "Status"
          },
          "predicted_price": {
            "type": "number",
            "title": "Predicted Price"
          },
          "currency": {
            "type": "string",
            "title": "Currency",
            "default": "USD"
          },
          "confidence_score": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confidence Score"
          },
          "model_name": {
            "type": "string",
            "title": "Model Name"
          },
          "model_version": {
            "type": "string",
            "title": "Model Version"
          },
          "feature_snapshot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Feature Snapshot Id"
          },
          "normalized_address": {
            "$ref": "#/components/schemas/NormalizedAddress"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "generated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Generated At"
          },
          "was_reused": {
            "type": "boolean",
            "title": "Was Reused",
            "default": false
          },
          "source_prediction_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Prediction Id"
          },
          "selected_feature_policy_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Feature Policy Name"
          },
          "selected_feature_policy_version": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selected Feature Policy Version"
          },
          "key_features": {
            "additionalProperties": true,
            "type": "object",
            "title": "Key Features",
            "description": "Key property features used for this prediction, e.g. bedrooms, living area sq ft, lot area, bathrooms, year built. Only populated features (non-null) are included."
          },
          "feature_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Source",
            "description": "Data source used to build the feature vector: 'census_context', 'census_context_with_backfill', 'heuristic', or 'fake'. Useful for auditing data quality in production."
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "request_id",
          "prediction_id",
          "correlation_id",
          "status",
          "predicted_price",
          "model_name",
          "model_version",
          "feature_snapshot_id",
          "normalized_address",
          "submitted_at",
          "generated_at"
        ],
        "title": "PredictionResponse"
      },
...
      "ScenarioBatchPipelineRequest": {
        "properties": {
          "scenario_ids": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Ids"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ScenarioBatchPipelineRequest",
        "description": "Request to run a subset (or all) registered scenarios through the pipeline."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}